// src/app/models/personal.model.ts
export interface PersonalState {
  firstName: string;
  lastName: string;
  birthDate: string;
  email: string;
  phone: string;
  address: string;
  city: string;
  zipCode: string;
}

export interface PersonalFormModel {
  firstName: string;
  lastName: string;
  birthDate: string;
  email: string;
  phone: string;
  address: string;
  city: string;
  zipCode: string;
}